Move mark cleanup to callback mode bug.
Active region isn't cleared properly in `READLINE_CALLBACKS` mode when
typing input: the mark is cleared on the second character typed, but
should be cleared on the first character typed.

The "keep mark active" logic was in a place that's reached for all
charloop inputs.  But in `READLINE_CALLBACKS` mode, it's not reached for
some input execution paths.  It looks like it belongs in
`_rl_internal_char_cleanup()`, which appears to exist for exactly this
purpose.

REPRO:
1. Use `READLINE_CALLBACKS` mode.
2. Start a fresh readline input line.
3. Type `abc`.
4. Invoke `beginning-of-line` command to move `rl_point` to beginning of
line.
5. Invoke `exchange-point-and-mark` command.
6. At this point the input line should display `abc` in reverse video
with `rl_point` at the end of the input line.
7. Type `d`.

EXPECTED: Input line should display `abcd` in normal video.
ACTUAL: Input line displays `abcd` in reverse video.  At that point,
typing `e` displays `abcde` in normal video.