Add conditional support for backslash as a path separator.
`rl_backslash_path_sep` defaults to zero, which supports only slash as a
path separator.  Set to non-zero to add backslash as a path separator.

`rl_preferred_path_separator` defaults to `'/'`, which inserts slash as
the path separator when needed.  Set to `'\\'` to insert backslash
instead.

These could be combined, but are currently separate with the intent that
a host could want to support parsing backslash, but always insert slash.