Add `search-ignore-case` config variable.
This optionally makes history search case insensitive.
E.g. this makes a big difference on Windows systems where it's common to
have some filename characters in upper case because completion mimicked
the casing from the file system.  But users are often not aware of the
actual casing that was used in a command line, because it generally
doesn't matter on Windows.