Add `add-history` and `remove-history` bindable commands.
`add-history` adds the current line to history without returning it, and
clears the input line.  This is handy when you've invested a lot of
typing to construct an input line, but you need to do something else
first before accepting that input line.

`remove-history` can be used while on a history entry (even when
searching history), and removes the current history entry from history.

NOTE:  If a host is providing custom history management, then both of
these commands will likely require additional support from a host in
order for the commands to integrate with the host's custom history
implementation.  That means there may be backward compatibility issues
with this change for some hosts.

What is a good way to make the Readline library accommodate the
potential additional hosting requirements?