Add support for a right side prompt string.
A host can call `rl_set_rprompt()` to set a right side prompt.  If the input is a single line and there is room past the end, then the right
side prompt is displayed right-justified on the first input line.  If
the input gets too long, the right side prompt is hidden.  If the input
becomes short enough, the right side prompt is shown again.

The right side prompt can contain escape codes, which must be surrounded
by `\001` and `\002` just like the normal prompt.

For example, a host could use an environment variable to let scripts
specify a right side prompt string.